--- InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/annotation/ThrowWithMessage.kt
@@ -1,7 +1,0 @@
-package com.jun.instatistoryautomatorserver.annotation
-
-@Target(AnnotationTarget.FUNCTION)
-@Retention(value = AnnotationRetention.RUNTIME)
-annotation class ThrowWithMessage(
-    val message: String,
-)
