+++ InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/application/dto/InstaApiResponseDTO.kt
@@ -1,0 +1,12 @@
+package com.jun.instatistoryautomatorserver.application.dto
+
+data class InstaApiResponseDTO (
+    var `data`: List<InstaPostResponseDTO>,
+    var paging: Paging?,
+) {
+    data class Paging(
+        var cursors: Any?,
+        var previous: String?,
+        var next: String?,
+    )
+}
