--- InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/dto/TistoryRequestDTO.kt
@@ -1,9 +1,0 @@
-package com.jun.instatistoryautomatorserver.dto
-
-data class TistoryRequestDTO (
-    val title: String,
-    val category: String = "술",
-    val imageUrl: String,
-    val content: String,
-    val tags: List<String> = listOf(),
-)
