--- InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/exception/InstaException.kt
@@ -1,6 +1,0 @@
-package com.jun.instatistoryautomatorserver.exception
-
-class InstaException(
-    override val message: String? = null,
-    override val cause: Throwable? = null
-    ): RuntimeException()
