--- InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/exception/TistoryException.kt
@@ -1,6 +1,0 @@
-package com.jun.instatistoryautomatorserver.exception
-
-class TistoryException(
-    override val message: String,
-    override val cause: Throwable? = null
-    ): RuntimeException()
