+++ InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/global/annotation/ThrowWithMessage.kt
@@ -1,0 +1,7 @@
+package com.jun.instatistoryautomatorserver.global.annotation
+
+@Target(AnnotationTarget.FUNCTION)
+@Retention(value = AnnotationRetention.RUNTIME)
+annotation class ThrowWithMessage(
+    val message: String,
+)
