+++ InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/global/exception/InstaException.kt
@@ -1,0 +1,6 @@
+package com.jun.instatistoryautomatorserver.global.exception
+
+class InstaException(
+    override val message: String? = null,
+    override val cause: Throwable? = null
+    ): RuntimeException()
