+++ InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/global/exception/TistoryException.kt
@@ -1,0 +1,6 @@
+package com.jun.instatistoryautomatorserver.global.exception
+
+class TistoryException(
+    override val message: String? = null,
+    override val cause: Throwable? = null
+    ): RuntimeException()
