+++ InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/global/property/TistoryProperty.kt
@@ -1,0 +1,10 @@
+package com.jun.instatistoryautomatorserver.global.property
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "tistory")
+data class TistoryProperty(
+    val entryUrl: String,
+    val email: String,
+    val password: String,
+)
