+++ InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/global/property/WebDriverProperty.kt
@@ -1,0 +1,10 @@
+package com.jun.instatistoryautomatorserver.global.property
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "webdriver")
+data class WebDriverProperty(
+    val timeout: Long = 3,
+    val arguments: List<String> = listOf(),
+    val chromeDriverPath: String,
+)
