+++ InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/global/type/MediaType.kt
@@ -1,0 +1,8 @@
+package com.jun.instatistoryautomatorserver.global.type
+
+enum class MediaType {
+    IMAGE,
+    VIDEO,
+    CAROUSEL_ALBUM,
+}
+
