--- InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/property/InstaProperty.kt
@@ -1,12 +1,0 @@
-package com.jun.instatistoryautomatorserver.property
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "insta")
-data class InstaProperty (
-    val baseUrl: String,
-    val apiVersion: String,
-    val userId: String,
-    val userToken: String,
-    val mediaFields: List<String>,
-)
