--- InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/property/TistoryProperty.kt
@@ -1,10 +1,0 @@
-package com.jun.instatistoryautomatorserver.property
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "tistory")
-data class TistoryProperty(
-    val entryUrl: String,
-    val email: String,
-    val password: String,
-)
