--- InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/property/WebDriverProperty.kt
@@ -1,10 +1,0 @@
-package com.jun.instatistoryautomatorserver.property
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "webdriver")
-data class WebDriverProperty(
-    val timeout: Long = 3,
-    val arguments: List<String> = listOf(),
-    val chromeDriverPath: String,
-)
