--- InstaTistoryAutomatorServer/src/main/kotlin/com/jun/instatistoryautomatorserver/repository/InstaRepository.kt
@@ -1,8 +1,0 @@
-package com.jun.instatistoryautomatorserver.repository
-
-import com.jun.instatistoryautomatorserver.entity.InstaPost
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface InstaRepository: JpaRepository<InstaPost, String> {
-    fun findFirstByOrderByTimestampDesc(): InstaPost?
-}
