+++ InstaTistoryAutomatorServer/src/test/kotlin/com/jun/instatistoryautomatorserver/insta/InstaServiceTest.kt
@@ -1,10 +1,12 @@
 package com.jun.instatistoryautomatorserver.insta
 
 import com.jun.instatistoryautomatorserver.insta.InstaServiceTest.Companion.PORT
-import com.jun.instatistoryautomatorserver.service.InstaService
+import com.jun.instatistoryautomatorserver.application.model.InstaService
+import com.jun.instatistoryautomatorserver.global.exception.InstaException
 import kotlinx.coroutines.test.runTest
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.kotlin.given
 import org.springframework.boot.test.mock.mockito.SpyBean
@@ -20,7 +22,7 @@ class InstaServiceTest {
     lateinit var instaService: InstaService
 
     @Test
-    fun fetchInstaPost() = runTest {
+    fun `유효한 인스타 API로 데이터 얻기 성공`() = runTest {
         val testUrl = "http://localhost:$PORT/valid-insta-api-entry-with-next-page"
         given(instaService.getInitialUrl()).willReturn(testUrl)
         val instaPosts = instaService.fetchInstaPost()
@@ -32,6 +34,16 @@ class InstaServiceTest {
         assertThat(instaPosts[3].instaId).isEqualTo("4")
         assertThat(instaPosts[4].instaId).isEqualTo("5")
         assertThat(instaPosts[5].instaId).isEqualTo("6")
+    }
+
+    @Test
+    fun `유효하지 않은 인스타 API로 데이터 얻기 실패, 예외 전파`() = runTest {
+        val testUrl = "http://localhost:$PORT/valid-insta-api-entry-with-invalid-next-page"
+        given(instaService.getInitialUrl()).willReturn(testUrl)
+
+        assertThrows<InstaException> {
+            instaService.fetchInstaPost()
+        }
     }
 
     companion object {
