+++ InstaTistoryAutomatorServer/src/test/kotlin/com/jun/instatistoryautomatorserver/insta/InstaTestConfig.kt
@@ -1,9 +1,9 @@
 package com.jun.instatistoryautomatorserver.insta
 
-import com.jun.instatistoryautomatorserver.config.RetrofitConfig
-import com.jun.instatistoryautomatorserver.property.InstaProperty
-import com.jun.instatistoryautomatorserver.repository.InstaRepository
-import com.jun.instatistoryautomatorserver.service.InstaService
+import com.jun.instatistoryautomatorserver.global.config.RetrofitConfig
+import com.jun.instatistoryautomatorserver.global.property.InstaProperty
+import com.jun.instatistoryautomatorserver.adapter.out.db.InstaRepository
+import com.jun.instatistoryautomatorserver.application.model.InstaService
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.boot.test.mock.mockito.MockBean
 import org.springframework.context.annotation.Import
